[TX] Add remaining platform tx to caminoExecutor (spend / lock)
