+++ e2e_tests/info_nomock.test.ts
