+++ src/apis/platformvm/utxos.ts
@@ -9,7 +9,6 @@ import {
   AmountOutput,
   SelectOutputClass,
   TransferableOutput,
-  SECPOwnerOutput,
   ParseableOutput,
   StakeableLockOut,
   SECPTransferOutput
@@ -23,38 +22,27 @@ import {
 import { UnixNow } from "../../utils/helperfunctions"
 import { StandardUTXO, StandardUTXOSet } from "../../common/utxos"
 import { PlatformVMConstants } from "./constants"
-import { UnsignedTx } from "./tx"
-import { ExportTx } from "../platformvm/exporttx"
-import { ImportTx } from "../platformvm/importtx"
-import { BaseTx } from "../platformvm/basetx"
 import {
   StandardAssetAmountDestination,
   AssetAmount
 } from "../../common/assetamount"
 import { BaseInput } from "../../common/input"
 import { BaseOutput } from "../../common/output"
-import { AddDelegatorTx, AddValidatorTx } from "./validationtx"
-import { CreateSubnetTx } from "./createsubnettx"
 import { Serialization, SerializedEncoding } from "../../utils/serialization"
 import {
   UTXOError,
   AddressError,
-  InsufficientFundsError,
-  ThresholdError,
-  FeeAssetError,
-  TimeError
+  InsufficientFundsError
 } from "../../utils/errors"
-import { DefaultNetworkID } from "../../utils/constants"
-import Networks from "../../utils/networks"
-import { CreateChainTx } from "."
-import { GenesisData } from "../avm"
-import { AddSubnetValidatorTx } from "../platformvm/addsubnetvalidatortx"
+import { LockMode } from "./builder"
 
 /**
  * @ignore
  */
 const bintools: BinTools = BinTools.getInstance()
 const serialization: Serialization = Serialization.getInstance()
+
+const zeroBN = new BN(0)
 
 /**
  * Class for representing a single UTXO.
@@ -252,16 +240,17 @@ export class UTXOSet extends StandardUTXOSet<UTXO> {
     })
   }
 
-  getMinimumSpendable = (
+  getMinimumSpendable = async (
     aad: AssetAmountDestination,
-    asOf: BN = UnixNow(),
-    locktime: BN = new BN(0),
-    threshold: number = 1,
-    stakeable: boolean = false
-  ): Error => {
-    let utxoArray: UTXO[] = this.getConsumableUXTO(asOf, stakeable)
+    asOf: BN = zeroBN,
+    lockTime: BN = zeroBN,
+    lockMode: LockMode = "Unlocked"
+  ): Promise<Error> => {
+    if (asOf.isZero()) asOf = UnixNow()
+
+    let utxoArray: UTXO[] = this.getConsumableUXTO(asOf, lockMode == "Stake")
     let tmpUTXOArray: UTXO[] = []
-    if (stakeable) {
+    if (lockMode == "Stake") {
       // If this is a stakeable transaction then have StakeableLockOut come before SECPTransferOutput
       // so that users first stake locked tokens before staking unlocked tokens
       utxoArray.forEach((utxo: UTXO) => {
@@ -482,7 +471,7 @@ export class UTXOSet extends StandardUTXOSet<UTXO> {
           unlockedChange,
           aad.getChangeAddresses(),
           zero.clone(), // make sure that we don't lock the change output.
-          threshold
+          aad.getChangeAddressesThreshold()
         ) as AmountOutput
         const transferOutput: TransferableOutput = new TransferableOutput(
           assetID,
@@ -508,8 +497,8 @@ export class UTXOSet extends StandardUTXOSet<UTXO> {
         const newOutput: AmountOutput = new SECPTransferOutput(
           unlockedAmount,
           aad.getDestinations(),
-          locktime,
-          threshold
+          lockTime,
+          aad.getDestinationsThreshold()
         ) as AmountOutput
         const transferOutput: TransferableOutput = new TransferableOutput(
           assetID,
@@ -519,825 +508,5 @@ export class UTXOSet extends StandardUTXOSet<UTXO> {
       }
     })
     return undefined
-  }
-
-  /**
-   * Creates an [[UnsignedTx]] wrapping a [[BaseTx]]. For more granular control, you may create your own
-   * [[UnsignedTx]] wrapping a [[BaseTx]] manually (with their corresponding [[TransferableInput]]s and [[TransferableOutput]]s).
-   *
-   * @param networkID The number representing NetworkID of the node
-   * @param blockchainID The {@link https://github.com/feross/buffer|Buffer} representing the BlockchainID for the transaction
-   * @param amount The amount of the asset to be spent in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}.
-   * @param assetID {@link https://github.com/feross/buffer|Buffer} of the asset ID for the UTXO
-   * @param toAddresses The addresses to send the funds
-   * @param fromAddresses The addresses being used to send the funds from the UTXOs {@link https://github.com/feross/buffer|Buffer}
-   * @param changeAddresses Optional. The addresses that can spend the change remaining from the spent UTXOs. Default: toAddresses
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned. Default: assetID
-   * @param memo Optional. Contains arbitrary data, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param locktime Optional. The locktime field created in the resulting outputs
-   * @param threshold Optional. The number of signatures required to spend the funds in the resultant UTXO
-   *
-   * @returns An unsigned transaction created from the passed in parameters.
-   *
-   */
-  buildBaseTx = (
-    networkID: number,
-    blockchainID: Buffer,
-    amount: BN,
-    assetID: Buffer,
-    toAddresses: Buffer[],
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[] = undefined,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow(),
-    locktime: BN = new BN(0),
-    threshold: number = 1
-  ): UnsignedTx => {
-    if (threshold > toAddresses.length) {
-      /* istanbul ignore next */
-      throw new ThresholdError(
-        "Error - UTXOSet.buildBaseTx: threshold is greater than number of addresses"
-      )
-    }
-
-    if (typeof changeAddresses === "undefined") {
-      changeAddresses = toAddresses
-    }
-
-    if (typeof feeAssetID === "undefined") {
-      feeAssetID = assetID
-    }
-
-    const zero: BN = new BN(0)
-
-    if (amount.eq(zero)) {
-      return undefined
-    }
-
-    const aad: AssetAmountDestination = new AssetAmountDestination(
-      toAddresses,
-      fromAddresses,
-      changeAddresses
-    )
-    if (assetID.toString("hex") === feeAssetID.toString("hex")) {
-      aad.addAssetAmount(assetID, amount, fee)
-    } else {
-      aad.addAssetAmount(assetID, amount, zero)
-      if (this._feeCheck(fee, feeAssetID)) {
-        aad.addAssetAmount(feeAssetID, zero, fee)
-      }
-    }
-
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-
-    const minSpendableErr: Error = this.getMinimumSpendable(
-      aad,
-      asOf,
-      locktime,
-      threshold
-    )
-    if (typeof minSpendableErr === "undefined") {
-      ins = aad.getInputs()
-      outs = aad.getAllOutputs()
-    } else {
-      throw minSpendableErr
-    }
-
-    const baseTx: BaseTx = new BaseTx(networkID, blockchainID, outs, ins, memo)
-    return new UnsignedTx(baseTx)
-  }
-
-  /**
-   * Creates an unsigned ImportTx transaction.
-   *
-   * @param networkID The number representing NetworkID of the node
-   * @param blockchainID The {@link https://github.com/feross/buffer|Buffer} representing the BlockchainID for the transaction
-   * @param toAddresses The addresses to send the funds
-   * @param fromAddresses The addresses being used to send the funds from the UTXOs {@link https://github.com/feross/buffer|Buffer}
-   * @param changeAddresses Optional. The addresses that can spend the change remaining from the spent UTXOs. Default: toAddresses
-   * @param importIns An array of [[TransferableInput]]s being imported
-   * @param sourceChain A {@link https://github.com/feross/buffer|Buffer} for the chainid where the imports are coming from.
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}. Fee will come from the inputs first, if they can.
-   * @param feeAssetID Optional. The assetID of the fees being burned.
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param locktime Optional. The locktime field created in the resulting outputs
-   * @param threshold Optional. The number of signatures required to spend the funds in the resultant UTXO
-   * @returns An unsigned transaction created from the passed in parameters.
-   *
-   */
-  buildImportTx = (
-    networkID: number,
-    blockchainID: Buffer,
-    toAddresses: Buffer[],
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[],
-    atomics: UTXO[],
-    sourceChain: Buffer = undefined,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow(),
-    locktime: BN = new BN(0),
-    threshold: number = 1
-  ): UnsignedTx => {
-    const zero: BN = new BN(0)
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-    if (typeof fee === "undefined") {
-      fee = zero.clone()
-    }
-
-    const importIns: TransferableInput[] = []
-    let feepaid: BN = new BN(0)
-    let feeAssetStr: string = feeAssetID.toString("hex")
-    for (let i: number = 0; i < atomics.length; i++) {
-      const utxo: UTXO = atomics[`${i}`]
-      const assetID: Buffer = utxo.getAssetID()
-      const output: AmountOutput = utxo.getOutput() as AmountOutput
-      let amt: BN = output.getAmount().clone()
-
-      let infeeamount = amt.clone()
-      let assetStr: string = assetID.toString("hex")
-      if (
-        typeof feeAssetID !== "undefined" &&
-        fee.gt(zero) &&
-        feepaid.lt(fee) &&
-        assetStr === feeAssetStr
-      ) {
-        feepaid = feepaid.add(infeeamount)
-        if (feepaid.gte(fee)) {
-          infeeamount = feepaid.sub(fee)
-          feepaid = fee.clone()
-        } else {
-          infeeamount = zero.clone()
-        }
-      }
-
-      const txid: Buffer = utxo.getTxID()
-      const outputidx: Buffer = utxo.getOutputIdx()
-      const input: SECPTransferInput = new SECPTransferInput(amt)
-      const xferin: TransferableInput = new TransferableInput(
-        txid,
-        outputidx,
-        assetID,
-        input
-      )
-      const from: Buffer[] = output.getAddresses()
-      const spenders: Buffer[] = output.getSpenders(from, asOf)
-      for (let j: number = 0; j < spenders.length; j++) {
-        const idx: number = output.getAddressIdx(spenders[`${j}`])
-        if (idx === -1) {
-          /* istanbul ignore next */
-          throw new AddressError(
-            "Error - UTXOSet.buildImportTx: no such " +
-              `address in output: ${spenders[`${j}`]}`
-          )
-        }
-        xferin.getInput().addSignatureIdx(idx, spenders[`${j}`])
-      }
-      importIns.push(xferin)
-      //add extra outputs for each amount (calculated from the imported inputs), minus fees
-      if (infeeamount.gt(zero)) {
-        const spendout: AmountOutput = SelectOutputClass(
-          output.getOutputID(),
-          infeeamount,
-          toAddresses,
-          locktime,
-          threshold
-        ) as AmountOutput
-        const xferout: TransferableOutput = new TransferableOutput(
-          assetID,
-          spendout
-        )
-        outs.push(xferout)
-      }
-    }
-
-    // get remaining fees from the provided addresses
-    let feeRemaining: BN = fee.sub(feepaid)
-    if (feeRemaining.gt(zero) && this._feeCheck(feeRemaining, feeAssetID)) {
-      const aad: AssetAmountDestination = new AssetAmountDestination(
-        toAddresses,
-        fromAddresses,
-        changeAddresses
-      )
-      aad.addAssetAmount(feeAssetID, zero, feeRemaining)
-      const minSpendableErr: Error = this.getMinimumSpendable(
-        aad,
-        asOf,
-        locktime,
-        threshold
-      )
-      if (typeof minSpendableErr === "undefined") {
-        ins = aad.getInputs()
-        outs = aad.getAllOutputs()
-      } else {
-        throw minSpendableErr
-      }
-    }
-
-    const importTx: ImportTx = new ImportTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      sourceChain,
-      importIns
-    )
-    return new UnsignedTx(importTx)
-  }
-
-  /**
-   * Creates an unsigned ExportTx transaction.
-   *
-   * @param networkID The number representing NetworkID of the node
-   * @param blockchainID The {@link https://github.com/feross/buffer|Buffer} representing the BlockchainID for the transaction
-   * @param amount The amount being exported as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param avaxAssetID {@link https://github.com/feross/buffer|Buffer} of the asset ID for AVAX
-   * @param toAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who recieves the AVAX
-   * @param fromAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who owns the AVAX
-   * @param changeAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who gets the change leftover of the AVAX
-   * @param destinationChain Optional. A {@link https://github.com/feross/buffer|Buffer} for the chainid where to send the asset.
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned.
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param locktime Optional. The locktime field created in the resulting outputs
-   * @param threshold Optional. The number of signatures required to spend the funds in the resultant UTXO
-   *
-   * @returns An unsigned transaction created from the passed in parameters.
-   *
-   */
-  buildExportTx = (
-    networkID: number,
-    blockchainID: Buffer,
-    amount: BN,
-    avaxAssetID: Buffer, // TODO: rename this to amountAssetID
-    toAddresses: Buffer[],
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[] = undefined,
-    destinationChain: Buffer = undefined,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow(),
-    locktime: BN = new BN(0),
-    threshold: number = 1
-  ): UnsignedTx => {
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-    let exportouts: TransferableOutput[] = []
-
-    if (typeof changeAddresses === "undefined") {
-      changeAddresses = toAddresses
-    }
-
-    const zero: BN = new BN(0)
-
-    if (amount.eq(zero)) {
-      return undefined
-    }
-
-    if (typeof feeAssetID === "undefined") {
-      feeAssetID = avaxAssetID
-    } else if (feeAssetID.toString("hex") !== avaxAssetID.toString("hex")) {
-      /* istanbul ignore next */
-      throw new FeeAssetError(
-        "Error - UTXOSet.buildExportTx: " + `feeAssetID must match avaxAssetID`
-      )
-    }
-
-    if (typeof destinationChain === "undefined") {
-      destinationChain = bintools.cb58Decode(
-        Networks.getNetwork(networkID).X.blockchainID
-      )
-    }
-
-    const aad: AssetAmountDestination = new AssetAmountDestination(
-      toAddresses,
-      fromAddresses,
-      changeAddresses
-    )
-    if (avaxAssetID.toString("hex") === feeAssetID.toString("hex")) {
-      aad.addAssetAmount(avaxAssetID, amount, fee)
-    } else {
-      aad.addAssetAmount(avaxAssetID, amount, zero)
-      if (this._feeCheck(fee, feeAssetID)) {
-        aad.addAssetAmount(feeAssetID, zero, fee)
-      }
-    }
-
-    const minSpendableErr: Error = this.getMinimumSpendable(
-      aad,
-      asOf,
-      locktime,
-      threshold
-    )
-    if (typeof minSpendableErr === "undefined") {
-      ins = aad.getInputs()
-      outs = aad.getChangeOutputs()
-      exportouts = aad.getOutputs()
-    } else {
-      throw minSpendableErr
-    }
-
-    const exportTx: ExportTx = new ExportTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      destinationChain,
-      exportouts
-    )
-
-    return new UnsignedTx(exportTx)
-  }
-
-  /**
-   * Class representing an unsigned [[AddSubnetValidatorTx]] transaction.
-   *
-   * @param networkID Networkid, [[DefaultNetworkID]]
-   * @param blockchainID Blockchainid, default undefined
-   * @param fromAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who pays the fees in AVAX
-   * @param changeAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who gets the change leftover from the fee payment
-   * @param nodeID The node ID of the validator being added.
-   * @param startTime The Unix time when the validator starts validating the Primary Network.
-   * @param endTime The Unix time when the validator stops validating the Primary Network (and staked AVAX is returned).
-   * @param weight The amount of weight for this subnet validator.
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned.
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param subnetAuthCredentials Optional. An array of index and address to sign for each SubnetAuth.
-   *
-   * @returns An unsigned transaction created from the passed in parameters.
-   */
-  buildAddSubnetValidatorTx = (
-    networkID: number = DefaultNetworkID,
-    blockchainID: Buffer,
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[],
-    nodeID: Buffer,
-    startTime: BN,
-    endTime: BN,
-    weight: BN,
-    subnetID: string,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow(),
-    subnetAuthCredentials: [number, Buffer][] = []
-  ): UnsignedTx => {
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-
-    const zero: BN = new BN(0)
-    const now: BN = UnixNow()
-    if (startTime.lt(now) || endTime.lte(startTime)) {
-      throw new Error(
-        "UTXOSet.buildAddSubnetValidatorTx -- startTime must be in the future and endTime must come after startTime"
-      )
-    }
-
-    if (this._feeCheck(fee, feeAssetID)) {
-      const aad: AssetAmountDestination = new AssetAmountDestination(
-        fromAddresses,
-        fromAddresses,
-        changeAddresses
-      )
-      aad.addAssetAmount(feeAssetID, zero, fee)
-      const success: Error = this.getMinimumSpendable(
-        aad,
-        asOf,
-        undefined,
-        undefined,
-        true
-      )
-      if (typeof success === "undefined") {
-        ins = aad.getInputs()
-        outs = aad.getAllOutputs()
-      } else {
-        throw success
-      }
-    }
-
-    const addSubnetValidatorTx: AddSubnetValidatorTx = new AddSubnetValidatorTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      nodeID,
-      startTime,
-      endTime,
-      weight,
-      subnetID
-    )
-    subnetAuthCredentials.forEach(
-      (subnetAuthCredential: [number, Buffer]): void => {
-        addSubnetValidatorTx.addSignatureIdx(
-          subnetAuthCredential[0],
-          subnetAuthCredential[1]
-        )
-      }
-    )
-    return new UnsignedTx(addSubnetValidatorTx)
-  }
-
-  /**
-   * Class representing an unsigned [[AddDelegatorTx]] transaction.
-   *
-   * @param networkID Networkid, [[DefaultNetworkID]]
-   * @param blockchainID Blockchainid, default undefined
-   * @param avaxAssetID {@link https://github.com/feross/buffer|Buffer} of the asset ID for AVAX
-   * @param toAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} recieves the stake at the end of the staking period
-   * @param fromAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who pays the fees and the stake
-   * @param changeAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who gets the change leftover from the staking payment
-   * @param nodeID The node ID of the validator being added.
-   * @param startTime The Unix time when the validator starts validating the Primary Network.
-   * @param endTime The Unix time when the validator stops validating the Primary Network (and staked AVAX is returned).
-   * @param stakeAmount A {@link https://github.com/indutny/bn.js/|BN} for the amount of stake to be delegated in nAVAX.
-   * @param rewardLocktime The locktime field created in the resulting reward outputs
-   * @param rewardThreshold The number of signatures required to spend the funds in the resultant reward UTXO
-   * @param rewardAddresses The addresses the validator reward goes.
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned.
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param changeThreshold Optional. The number of signatures required to spend the funds in the change UTXO
-   *
-   * @returns An unsigned transaction created from the passed in parameters.
-   */
-  buildAddDelegatorTx = (
-    networkID: number = DefaultNetworkID,
-    blockchainID: Buffer,
-    avaxAssetID: Buffer,
-    toAddresses: Buffer[],
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[],
-    nodeID: Buffer,
-    startTime: BN,
-    endTime: BN,
-    stakeAmount: BN,
-    rewardLocktime: BN,
-    rewardThreshold: number,
-    rewardAddresses: Buffer[],
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow(),
-    changeThreshold: number = 1
-  ): UnsignedTx => {
-    if (rewardThreshold > rewardAddresses.length) {
-      /* istanbul ignore next */
-      throw new ThresholdError(
-        "Error - UTXOSet.buildAddDelegatorTx: reward threshold is greater than number of addresses"
-      )
-    }
-
-    if (typeof changeAddresses === "undefined") {
-      changeAddresses = toAddresses
-    }
-
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-    let stakeOuts: TransferableOutput[] = []
-
-    const zero: BN = new BN(0)
-    const now: BN = UnixNow()
-    if (startTime.lt(now) || endTime.lte(startTime)) {
-      throw new TimeError(
-        "UTXOSet.buildAddDelegatorTx -- startTime must be in the future and endTime must come after startTime"
-      )
-    }
-
-    const aad: AssetAmountDestination = new AssetAmountDestination(
-      toAddresses,
-      fromAddresses,
-      changeAddresses
-    )
-    if (avaxAssetID.toString("hex") === feeAssetID.toString("hex")) {
-      aad.addAssetAmount(avaxAssetID, stakeAmount, fee)
-    } else {
-      aad.addAssetAmount(avaxAssetID, stakeAmount, zero)
-      if (this._feeCheck(fee, feeAssetID)) {
-        aad.addAssetAmount(feeAssetID, zero, fee)
-      }
-    }
-
-    const minSpendableErr: Error = this.getMinimumSpendable(
-      aad,
-      asOf,
-      undefined,
-      changeThreshold,
-      true
-    )
-    if (typeof minSpendableErr === "undefined") {
-      ins = aad.getInputs()
-      outs = aad.getChangeOutputs()
-      stakeOuts = aad.getOutputs()
-    } else {
-      throw minSpendableErr
-    }
-
-    const rewardOutputOwners: SECPOwnerOutput = new SECPOwnerOutput(
-      rewardAddresses,
-      rewardLocktime,
-      rewardThreshold
-    )
-
-    const UTx: AddDelegatorTx = new AddDelegatorTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      nodeID,
-      startTime,
-      endTime,
-      stakeAmount,
-      stakeOuts,
-      new ParseableOutput(rewardOutputOwners)
-    )
-    return new UnsignedTx(UTx)
-  }
-
-  /**
-   * Class representing an unsigned [[AddValidatorTx]] transaction.
-   *
-   * @param networkID NetworkID, [[DefaultNetworkID]]
-   * @param blockchainID BlockchainID, default undefined
-   * @param avaxAssetID {@link https://github.com/feross/buffer|Buffer} of the asset ID for AVAX
-   * @param toAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} recieves the stake at the end of the staking period
-   * @param fromAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who pays the fees and the stake
-   * @param changeAddresses An array of addresses as {@link https://github.com/feross/buffer|Buffer} who gets the change leftover from the staking payment
-   * @param nodeID The node ID of the validator being added.
-   * @param startTime The Unix time when the validator starts validating the Primary Network.
-   * @param endTime The Unix time when the validator stops validating the Primary Network (and staked AVAX is returned).
-   * @param stakeAmount A {@link https://github.com/indutny/bn.js/|BN} for the amount of stake to be delegated in nAVAX.
-   * @param rewardLocktime The locktime field created in the resulting reward outputs
-   * @param rewardThreshold The number of signatures required to spend the funds in the resultant reward UTXO
-   * @param rewardAddresses The addresses the validator reward goes.
-   * @param delegationFee A number for the percentage of reward to be given to the validator when someone delegates to them. Must be between 0 and 100.
-   * @param minStake A {@link https://github.com/indutny/bn.js/|BN} representing the minimum stake required to validate on this network.
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned.
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   *
-   * @returns An unsigned transaction created from the passed in parameters.
-   */
-  buildAddValidatorTx = (
-    networkID: number = DefaultNetworkID,
-    blockchainID: Buffer,
-    avaxAssetID: Buffer,
-    toAddresses: Buffer[],
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[],
-    nodeID: Buffer,
-    startTime: BN,
-    endTime: BN,
-    stakeAmount: BN,
-    rewardLocktime: BN,
-    rewardThreshold: number,
-    rewardAddresses: Buffer[],
-    delegationFee: number,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow()
-  ): UnsignedTx => {
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-    let stakeOuts: TransferableOutput[] = []
-
-    const zero: BN = new BN(0)
-    const now: BN = UnixNow()
-    if (startTime.lt(now) || endTime.lte(startTime)) {
-      throw new TimeError(
-        "UTXOSet.buildAddValidatorTx -- startTime must be in the future and endTime must come after startTime"
-      )
-    }
-
-    if (delegationFee > 100 || delegationFee < 0) {
-      throw new TimeError(
-        "UTXOSet.buildAddValidatorTx -- startTime must be in the range of 0 to 100, inclusively"
-      )
-    }
-
-    const aad: AssetAmountDestination = new AssetAmountDestination(
-      toAddresses,
-      fromAddresses,
-      changeAddresses
-    )
-    if (avaxAssetID.toString("hex") === feeAssetID.toString("hex")) {
-      aad.addAssetAmount(avaxAssetID, stakeAmount, fee)
-    } else {
-      aad.addAssetAmount(avaxAssetID, stakeAmount, zero)
-      if (this._feeCheck(fee, feeAssetID)) {
-        aad.addAssetAmount(feeAssetID, zero, fee)
-      }
-    }
-
-    const minSpendableErr: Error = this.getMinimumSpendable(
-      aad,
-      asOf,
-      undefined,
-      undefined,
-      true
-    )
-    if (typeof minSpendableErr === "undefined") {
-      ins = aad.getInputs()
-      outs = aad.getChangeOutputs()
-      stakeOuts = aad.getOutputs()
-    } else {
-      throw minSpendableErr
-    }
-
-    const rewardOutputOwners: SECPOwnerOutput = new SECPOwnerOutput(
-      rewardAddresses,
-      rewardLocktime,
-      rewardThreshold
-    )
-
-    const UTx: AddValidatorTx = new AddValidatorTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      nodeID,
-      startTime,
-      endTime,
-      stakeAmount,
-      stakeOuts,
-      new ParseableOutput(rewardOutputOwners),
-      delegationFee
-    )
-    return new UnsignedTx(UTx)
-  }
-
-  /**
-   * Class representing an unsigned [[CreateSubnetTx]] transaction.
-   *
-   * @param networkID Networkid, [[DefaultNetworkID]]
-   * @param blockchainID Blockchainid, default undefined
-   * @param fromAddresses The addresses being used to send the funds from the UTXOs {@link https://github.com/feross/buffer|Buffer}
-   * @param changeAddresses The addresses that can spend the change remaining from the spent UTXOs.
-   * @param subnetOwnerAddresses An array of {@link https://github.com/feross/buffer|Buffer} for the addresses to add to a subnet
-   * @param subnetOwnerThreshold The number of owners's signatures required to add a validator to the network
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   *
-   * @returns An unsigned transaction created from the passed in parameters.
-   */
-  buildCreateSubnetTx = (
-    networkID: number = DefaultNetworkID,
-    blockchainID: Buffer,
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[],
-    subnetOwnerAddresses: Buffer[],
-    subnetOwnerThreshold: number,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow()
-  ): UnsignedTx => {
-    const zero: BN = new BN(0)
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-
-    if (this._feeCheck(fee, feeAssetID)) {
-      const aad: AssetAmountDestination = new AssetAmountDestination(
-        fromAddresses,
-        fromAddresses,
-        changeAddresses
-      )
-      aad.addAssetAmount(feeAssetID, zero, fee)
-      const minSpendableErr: Error = this.getMinimumSpendable(
-        aad,
-        asOf,
-        undefined,
-        undefined
-      )
-      if (typeof minSpendableErr === "undefined") {
-        ins = aad.getInputs()
-        outs = aad.getAllOutputs()
-      } else {
-        throw minSpendableErr
-      }
-    }
-
-    const locktime: BN = new BN(0)
-    const subnetOwners: SECPOwnerOutput = new SECPOwnerOutput(
-      subnetOwnerAddresses,
-      locktime,
-      subnetOwnerThreshold
-    )
-    const createSubnetTx: CreateSubnetTx = new CreateSubnetTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      subnetOwners
-    )
-
-    return new UnsignedTx(createSubnetTx)
-  }
-
-  /**
-   * Build an unsigned [[CreateChainTx]].
-   *
-   * @param networkID Networkid, [[DefaultNetworkID]]
-   * @param blockchainID Blockchainid, default undefined
-   * @param fromAddresses The addresses being used to send the funds from the UTXOs {@link https://github.com/feross/buffer|Buffer}
-   * @param changeAddresses The addresses that can spend the change remaining from the spent UTXOs.
-   * @param subnetID Optional ID of the Subnet that validates this blockchain
-   * @param chainName Optional A human readable name for the chain; need not be unique
-   * @param vmID Optional ID of the VM running on the new chain
-   * @param fxIDs Optional IDs of the feature extensions running on the new chain
-   * @param genesisData Optional Byte representation of genesis state of the new chain
-   * @param fee Optional. The amount of fees to burn in its smallest denomination, represented as {@link https://github.com/indutny/bn.js/|BN}
-   * @param feeAssetID Optional. The assetID of the fees being burned
-   * @param memo Optional contains arbitrary bytes, up to 256 bytes
-   * @param asOf Optional. The timestamp to verify the transaction against as a {@link https://github.com/indutny/bn.js/|BN}
-   * @param subnetAuthCredentials Optional. An array of index and address to sign for each SubnetAuth.
-   *
-   * @returns An unsigned CreateChainTx created from the passed in parameters.
-   */
-  buildCreateChainTx = (
-    networkID: number = DefaultNetworkID,
-    blockchainID: Buffer,
-    fromAddresses: Buffer[],
-    changeAddresses: Buffer[],
-    subnetID: string | Buffer = undefined,
-    chainName: string = undefined,
-    vmID: string = undefined,
-    fxIDs: string[] = undefined,
-    genesisData: string | GenesisData = undefined,
-    fee: BN = undefined,
-    feeAssetID: Buffer = undefined,
-    memo: Buffer = undefined,
-    asOf: BN = UnixNow(),
-    subnetAuthCredentials: [number, Buffer][] = []
-  ): UnsignedTx => {
-    const zero: BN = new BN(0)
-    let ins: TransferableInput[] = []
-    let outs: TransferableOutput[] = []
-
-    if (this._feeCheck(fee, feeAssetID)) {
-      const aad: AssetAmountDestination = new AssetAmountDestination(
-        fromAddresses,
-        fromAddresses,
-        changeAddresses
-      )
-      aad.addAssetAmount(feeAssetID, zero, fee)
-      const minSpendableErr: Error = this.getMinimumSpendable(
-        aad,
-        asOf,
-        undefined,
-        undefined
-      )
-      if (typeof minSpendableErr === "undefined") {
-        ins = aad.getInputs()
-        outs = aad.getAllOutputs()
-      } else {
-        throw minSpendableErr
-      }
-    }
-
-    const createChainTx: CreateChainTx = new CreateChainTx(
-      networkID,
-      blockchainID,
-      outs,
-      ins,
-      memo,
-      subnetID,
-      chainName,
-      vmID,
-      fxIDs,
-      genesisData
-    )
-    subnetAuthCredentials.forEach(
-      (subnetAuthCredential: [number, Buffer]): void => {
-        createChainTx.addSignatureIdx(
-          subnetAuthCredential[0],
-          subnetAuthCredential[1]
-        )
-      }
-    )
-
-    return new UnsignedTx(createChainTx)
   }
 }
