+++ tests/apis/platformvm/tx.test.ts
@@ -22,6 +22,7 @@ import { BaseTx } from "../../../src/apis/platformvm/basetx"
 import { ImportTx } from "../../../src/apis/platformvm/importtx"
 import { ExportTx } from "../../../src/apis/platformvm/exporttx"
 import { DefaultPlatformChainID } from "src/utils"
+import { Builder } from "../../../src/apis/platformvm/builder"
 
 describe("Transactions", (): void => {
   /**
@@ -31,6 +32,7 @@ describe("Transactions", (): void => {
 
   const networkID: number = 12345
   let set: UTXOSet
+  let builder: Builder
   let keymgr1: KeyChain
   let keymgr2: KeyChain
   let keymgr3: KeyChain
@@ -93,6 +95,7 @@ describe("Transactions", (): void => {
 
   beforeEach((): void => {
     set = new UTXOSet()
+    builder = new Builder(set, false)
     keymgr1 = new KeyChain(avalanche.getHRP(), alias)
     keymgr2 = new KeyChain(avalanche.getHRP(), alias)
     keymgr3 = new KeyChain(avalanche.getHRP(), alias)
@@ -405,9 +408,10 @@ describe("Transactions", (): void => {
     expect(txunew.toString()).toBe(txu.toString())
   })
 
-  test("Creation UnsignedTx Check Amount", (): void => {
-    expect((): void => {
-      set.buildBaseTx(
+  test("Creation UnsignedTx Check Amount", (): Promise<void | UnsignedTx> => {
+    expect.assertions(1)
+    return builder
+      .buildBaseTx(
         netid,
         blockchainID,
         new BN(amnt * 1000),
@@ -416,7 +420,7 @@ describe("Transactions", (): void => {
         addrs1,
         addrs1
       )
-    }).toThrow()
+      .catch((e) => expect(e).toBeDefined())
   })
 
   test("Creation ImportTx", (): void => {
@@ -493,8 +497,8 @@ describe("Transactions", (): void => {
     expect(exportTx.getExportOutputs().length).toBe(exportOuts.length)
   })
 
-  test("Creation Tx1 with asof, locktime, threshold", (): void => {
-    const txu: UnsignedTx = set.buildBaseTx(
+  test("Creation Tx1 with asof, locktime, threshold", async (): Promise<void> => {
+    const txu: UnsignedTx = await builder.buildBaseTx(
       netid,
       blockchainID,
       new BN(9000),
@@ -516,8 +520,8 @@ describe("Transactions", (): void => {
     expect(tx2.toBuffer().toString("hex")).toBe(tx.toBuffer().toString("hex"))
     expect(tx2.toString()).toBe(tx.toString())
   })
-  test("Creation Tx2 without asof, locktime, threshold", (): void => {
-    const txu: UnsignedTx = set.buildBaseTx(
+  test("Creation Tx2 without asof, locktime, threshold", async (): Promise<void> => {
+    const txu: UnsignedTx = await builder.buildBaseTx(
       netid,
       blockchainID,
       new BN(9000),
@@ -533,8 +537,8 @@ describe("Transactions", (): void => {
     expect(tx2.toString()).toBe(tx.toString())
   })
 
-  test("Creation Tx4 using ImportTx", (): void => {
-    const txu: UnsignedTx = set.buildImportTx(
+  test("Creation Tx4 using ImportTx", async (): Promise<void> => {
+    const txu: UnsignedTx = await builder.buildImportTx(
       netid,
       blockchainID,
       addrs3,
@@ -553,17 +557,17 @@ describe("Transactions", (): void => {
     expect(tx2.toBuffer().toString("hex")).toBe(tx.toBuffer().toString("hex"))
   })
 
-  test("Creation Tx5 using ExportTx", (): void => {
-    const txu: UnsignedTx = set.buildExportTx(
+  test("Creation Tx5 using ExportTx", async (): Promise<void> => {
+    const txu: UnsignedTx = await builder.buildExportTx(
       netid,
       blockchainID,
       new BN(90),
       avaxAssetID,
       addrs3,
       addrs1,
-      addrs2,
       bintools.cb58Decode(DefaultPlatformChainID),
-      undefined,
+      addrs2,
+      new BN(0),
       undefined,
       new UTF8Payload("hello world").getPayload(),
       UnixNow()
